Move print functions to `print` package
